Moves HoG to Avatar.java
